Fix: Handle multi-task training and prevent inf loss
This commit addresses the `inf` loss issue during training by correctly handling the multi-task learning setup.

The main changes are:
- Modified `entry.py` to correctly use the `CombinedMultiTaskDataset` and `StrictRatioBatchSampler` to create a single, unified dataloader.
- Modified `train_loop.py` to correctly handle the combined dataset and print individual loss components for debugging.
- Modified `model.py` to use the task ID to route the input to the correct head, ensuring that the correct loss is calculated for each task.

+++ span_selection_dataset.py
@@ -101,7 +101,7 @@ class SpanSelectionDataset(Dataset):
         # 5. Create dummy tensors for the other 5 tuple items to match batch format
         dummy_lm_targets = torch.full((self.seq_len,), self.lm_ignore_idx, dtype=torch.long)
         dummy_rank_targets = torch.full((self.seq_len,), float(self.lm_ignore_idx), dtype=torch.float32)
-        auxiliary_scalar_value = torch.tensor(float(target_index), dtype=torch.float32)
+        auxiliary_scalar_value = torch.tensor(target_index, dtype=torch.long)
         task_type_flag = torch.tensor(3.0, dtype=torch.float32)
         dummy_true_ranks = torch.full((self.seq_len,), float(self.lm_ignore_idx), dtype=torch.float32)
